Inclusão do teste adicional no login de mantenedores

+++ adminroot/removeLicao.php
@@ -32,7 +32,7 @@ try{
     
     Lib\DAO::setFilePathConfig("../assets/conexao.ini");
 
-    if (isset($_SESSION["logado"]) && isset($_SESSION["nome"])){
+    if (isset($_SESSION["logado"]) && isset($_SESSION["managerRoot"]) && isset($_SESSION["nome"])){
         
         if (isset($_GET["id"]) && Lib\DAOUtilis::isIntString($_GET["id"])){
             Lib\DAO::removeById($_GET["id"], "Licoes");
